# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/about/update": {
            "post": {
                "description": "Update About",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update About",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "User request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateAbout"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseOK"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/user/all": {
            "get": {
                "description": "Get Users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get Users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetUsersRequest"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "description": "Get User Info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get User Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "Login User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Login User",
                "parameters": [
                    {
                        "description": "User request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseOKWithJWT"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/user/quest": {
            "get": {
                "description": "Get User Quest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get User Quest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AllQuestUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "Registration User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Registration User",
                "parameters": [
                    {
                        "description": "User request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseOKWithJWT"
                        }
                    },
                    "400": {
                        "description": "Bad request (client fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal error (server fault)",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseFailure"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "base.Blame": {
            "type": "string",
            "enum": [
                "User",
                "Postgres",
                "Server",
                "Unknown",
                "Neo4j"
            ],
            "x-enum-varnames": [
                "BlameUser",
                "BlamePostgres",
                "BlameServer",
                "BlameUnknown",
                "BlameNeo4j"
            ]
        },
        "base.ResponseFailure": {
            "type": "object",
            "properties": {
                "blame": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/base.Blame"
                        }
                    ],
                    "example": "Guilty System"
                },
                "message": {
                    "type": "string",
                    "example": "error occurred"
                },
                "status": {
                    "type": "string",
                    "example": "Error"
                },
                "trackingID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                }
            }
        },
        "base.ResponseOK": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "trackingID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                }
            }
        },
...
        "base.ResponseOKWithJWT": {
            "type": "object",
            "properties": {
                "GUID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                },
                "JWT": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "trackingID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                }
            }
        },
        "model.AboutObject": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fio": {
                    "type": "string"
                }
            }
        },
...
        "model.AllQuestUserResponse": {
            "type": "object",
            "properties": {
                "divQuests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserQuestObject"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "trackingID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                }
            }
        },
...
        "model.CreateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.DivisionObject": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.GetUsersRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "trackingID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserObject"
                    }
                }
            }
        },
        "model.LoginUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.OverQuestObject": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StageObject"
                    }
                }
            }
        },
...
        "model.StageObject": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.UpdateAbout": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fio": {
                    "type": "string"
                }
            }
        },
        "model.UserInfoResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "trackingID": {
                    "type": "string",
                    "example": "12345678-1234-1234-1234-000000000000"
                },
                "user": {
                    "$ref": "#/definitions/model.UserObject"
                }
            }
        },
        "model.UserObject": {
            "type": "object",
            "properties": {
                "about": {
                    "$ref": "#/definitions/model.AboutObject"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "model.UserQuestObject": {
            "type": "object",
            "properties": {
                "division": {
                    "$ref": "#/definitions/model.DivisionObject"
                },
                "quests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OverQuestObject"
                    }
                }
            }
        }
    }
}